+++ broker/http/options.go
@@ -5,7 +5,10 @@ import (
 	"net/http"
 
 	"github.com/micro/go-micro/broker"
+	"github.com/micro/go-micro/registry"
 )
+
+type registryKey struct{}
 
 // Handle registers the handler for the given pattern.
 func Handle(pattern string, handler http.Handler) broker.Option {
@@ -19,5 +22,12 @@ func Handle(pattern string, handler http.Handler) broker.Option {
 		}
 		handlers[pattern] = handler
 		o.Context = context.WithValue(o.Context, "http_handlers", handlers)
+	}
+}
+
+// Registry sets the broker internal registry
+func Registry(r registry.Registry) broker.Option {
+	return func(o *broker.Options) {
+		o.Context = context.WithValue(o.Context, registryKey{}, r)
 	}
 }
